feat: add stimulus | event listener for element create with AJAX
