+++ assets/js/user.js
@@ -2,21 +2,6 @@ import axios from 'axios';
 
 const follow = document.querySelectorAll('[data-follow-user]');
 
-if(follow){
-    follow.forEach((element=>{
-        element.addEventListener('click', () =>  {
-            
-            let id = element.value;
-            axios.get(`/user/follow/${id}`);
-            if(element.children[0].classList.contains("far")){
-                element.children[0].classList.replace('far','fa' );
-            }else{
-                element.children[0].classList.replace('fa','far' );
-            }
-        });
-    }))
-}
-
 const styleUser = document.querySelectorAll('[data-style-user]');
 if(styleUser){
     styleUser.forEach((element=>{
@@ -40,17 +25,6 @@ if(deleteUser){
         });
     }))
 }
-
-// const deleteGood = document.querySelectorAll('[data-delete]');
-// if(deleteGood){
-//     deleteGood.forEach((element=>{
-//         element.addEventListener('click', () =>  {
-    
-//             axios.delete(`/delete`);
- 
-//         });
-//     }))
-// }
 
 const retour = document.querySelectorAll('[data-retour]');
 if(retour){
