+++ templates/subject/index.html.twig
@@ -48,11 +48,19 @@
         
 
         {# clone commentaire #}
-        <div class="comment3 modele" id="modele-comment">           
+        <div 
+            class="comment3 modele" 
+            id="modele-comment">           
             <div class="jaime" >
                 <span class="floatRight">
                     <span  id="note0">(0)</span> 
-                    <button id="j0" data_note="0/" class="btn-vierge jaimee" type="button" name="jaime" >
+                    <button 
+                        id="j0" 
+                        data_note="0/"
+                        class="btn-vierge jaimee" 
+                        type="button" 
+                        name="jaime" 
+                        >
                         <i data_id="0" class="far fa-thumbs-up"></i>
                     </button>
                 </span>
